app/requirements.txt
python-dotenv>=1.0.1
streamlit>=1.31.0
pickpod>=1.0.5
